requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
gunicorn==20.0.4
pandas==1.2.3